Manual subscription to response topic for autopaho RPC handler
To be able to continue using the same RPC handler across several
autopaho (re)connects means we must re-subscribe to the response topic
every time we come up online again. The simplest way to achieve this
is to make the handling of this subscription and also registering the
response topic handler in the router completely manual.